Improve the ide diagnostics trait API
